requirements.txt
streamlit
google-api-python-client
python-dotenv
pandas
sqlalchemy
pymysql
psycopg2
sqlalchemy
psycopg2-binary
